// src/app/core/interfaces/sorting.ts
export interface Sorting {
  id: 'asc' | 'desc' | 'none';
  username: 'asc' | 'desc' | 'none';
  password: 'asc' | 'desc' | 'none';
  name: 'asc' | 'desc' | 'none';
  surname: 'asc' | 'desc' | 'none';
  email: 'asc' | 'desc' | 'none';
  registrationDate: 'asc' | 'desc' | 'none';
  isActive: 'asc' | 'desc' | 'none';
}